+++ src/mongodb/collection/account_token/account_token_find_account.rs
@@ -12,18 +12,18 @@ struct Filter {
     #[serde(rename = "_id")]
     #[serde(serialize_with = "serialize_hex_string_as_object_id")]
     id: String,
-    issue: usize,
-    expiry: usize,
+    issue: u64,
+    expiry: u64,
 }
 
 #[async_trait]
 pub trait FindAccount {
-    async fn find_account(&self, tok: String, iat: usize, exp: usize) -> Result;
+    async fn find_account(&self, tok: String, iat: u64, exp: u64) -> Result;
 }
 
 #[async_trait]
 impl FindAccount for Collection<AccountToken> {
-    async fn find_account(&self, tok: String, iat: usize, exp: usize) -> Result {
+    async fn find_account(&self, tok: String, iat: u64, exp: u64) -> Result {
         let filter = Filter {
             id: tok, issue: iat, expiry: exp
         };
